+++ plugins/circleci/src/pages/BuildWithStepsPage/index.ts
@@ -13,4 +13,4 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-export { default as DetailedViewPage } from './DetailedViewPage';
+export { default as DetailedViewPage } from './BuildWithStepsPage';
